ADD: GetTotalIncome

+++ src/main/java/se/johan/kvitt/event/dto/request/EventCreateEventRequestDTO.java
@@ -4,13 +4,15 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
+
 
 public record EventCreateEventRequestDTO(
         @Size(max = 255, message = "Titeln får inte vara längre än 255 tecken")
         @NotBlank(message = "Titeln får inte vara tomt")
         String title,
         @NotNull
-        double amount,
+        BigDecimal amount,
         @NotNull
         boolean expense,
         @Size(max = 255, message = "ID får inte vara längre än 255 tecken")
