+++ src/main/java/se/johan/kvitt/event/model/Event.java
@@ -4,13 +4,14 @@ package se.johan.kvitt.event.model;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 @Document(collection = "kvittCollection")
 public record Event(
     @Id
     String id,
     String title,
-    double amount,
+    BigDecimal amount,
     boolean expense,
     LocalDateTime dateTime,
     String kvittUserId
